+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-	"python.pythonPath": "C:\\ProgramData\\Anaconda3\\envs\\sentinello\\python.exe"
+	"python.pythonPath": "e:\\Studia\\Praktyki\\Python\\venv\\Scripts\\python.exe"
 }
